+++ libs/types/src/index.ts
@@ -1,1 +1,2 @@
 export * from './common'
+export * from './eip6963-types'